Moved components to feature folder
